Fixes a critical bug and adds a new feature
This commit
1)	fixes a bug in the original repo
2)	allows the user to generate reproducible results
3)	extends the code from stateless to contextual bandit
4)	separates training and back-testing sections of the code.

For the details of the critical bug and my solution,
see the issue opened in the original repo.
https://github.com/wai-i/Pair-Trading-Reinforcement-Learning/issues/8.

In the original repo, random number generators were not
properly seeded, so the results were not reproducible.
This commit fixed that problem.

Extending from stateless to contextual bandit allows
the agent to adjust the trading parameters according
to the environment's current state.
In the committed version, the environment can have
only two states: positive or negative (depending on
whether the spread is above or below the historical value).
In the tests I performed before making this commit,
the optimal parameters learned by the agent with binary
environment representation were virtually the same as stateless.
In future versions, the environment should be
allowed to have more states.

I separated training and back-testing parts because it
allows one to isolate potential problems better.
At the end of the training, the TensorFlow graph is saved
and then restored in back-testing.

Also, price data for the top 50 indices in each of the
top three S&P500 sectors are downloaded (from Tiingo)
into the working directory.